Runcomm (#61)
* wip

* deviceloop as class

* Added one command, terminate and runclient.py to test it

* Run in simulate mode

* Status working

* Working load

* Alternative webserver

* command,data tuple return

* Updated test

* Separate out status set

* Getting test to work

* Make test not fail if OK

* wip

* Initial web server working

* Updated golden files

* Clean test same as master

* Understanding forms

* Get requirements right

* Command radiobuttons working

+++ src/communicate/testnetsock.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
-timeout 20 python ./mocksrv.py &
+timeout 60 python ./mocksrv.py &
 sleep 1
 python ./mockclient.py
 
-pkill -9 "python ./mocksrv.py"
+# Check if mocksrv.py is running, if so forcefully kill it
+if $(pgrep mocksrv.py  &> /dev/null)
+then 
+  pkill -9 "python ./mocksrv.py"
+  exit 1
+fi
